requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
google-generativeai==0.3.2
SpeechRecognition==3.10.1
deep-translator==1.11.4
gTTS==2.5.0
playsound==1.3.0
